UML diagrams
Uploading these, because they was never uploaded. The class diagram is a old version with some new additions like the DB handler and Contract enum. An updated version will be created to give an accuraet view of our classes.

The form uml is a view of all the forms and how they interact with each other and classes.